main() CLI
It will attempt to login until there are no errors. Timeout is quite
generous because sometimes it takes a while to respond. Retry prevents us
from hammering the service too much.

It's more difficult to test this because we'd have to signal a way to stop
the loop when the test is complete. So I'm not going to for the moment.